Changes to matrix impl (WIP) tests
