+++ math/shared/src/main/scala/io/youi/math/Matrix3ArrayPool.scala
@@ -1,4 +1,4 @@
-package io.youi.math.test
+package io.youi.math
 
 import io.youi.util.ObjectPool
