+++ math/shared/src/main/scala/io/youi/math/package.scala
@@ -3,19 +3,7 @@ package io.youi
 import scala.math._
 
 package object math {
-
-  /**
-   * Create a new array backing 3d matrices of the correct size
-   */
-  def matrix3Array: Array[Double] = new Array[Double](9)
-
-  def identityArray: Array[Double] = {
-    val id = matrix3Array
-    id(Matrix3.M00) = 1
-    id(Matrix3.M11) = 1
-    id(Matrix3.M22) = 1
-    id
-  }
+  val precision: Double = 0.00001
 
   /**
    * Degrees wrapper
@@ -32,14 +20,22 @@ package object math {
    *
    * @param d the double value to convert
    */
-  implicit class AngleConvert(val d: Double) extends AnyVal {
+  implicit class DoubleOps(val d: Double) extends AnyVal {
     def rad: Radians = Radians(d)
 
     def deg: Degrees = Degrees(d)
+
+    def <=>(other: Double): Boolean = tolerance(d, other)
   }
 
   implicit class DegToRad(val deg: Degrees) {
     def toRad: Radians = Radians(toRadians(deg.value))
   }
+
+  def tolerantEquals(precision: Double)(d1: Double, d2: Double): Boolean = {
+    Math.abs(d1 - d2) <= precision
+  }
+
+  lazy val tolerance: (Double, Double) => Boolean = tolerantEquals(precision)
 
 }
